project_11_k8s_deployment/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
prometheus-client>=0.19.0
structlog>=24.1.0
redis>=5.0.0
aiofiles>=23.2.0
pyyaml>=6.0
jsonschema>=4.21.0
numpy>=1.26.0
scipy>=1.12.0
python-dotenv>=1.0.0
httpx>=0.26.0
tenacity>=8.2.0
langgraph-checkpoint-postgres>=1.0.0
langgraph-checkpoint-sqlite>=1.0.0
